perf: move escape html to utils
